Updated PCRE to 8.40
This updates PCRE to 8.40.  The source code was downloaded from
https://sourceforge.net/projects/pcre/files/pcre/8.40/ as this was the
unofficial redirect from www.pcre.org.  The svn command provided on the
main site did not work as there was no response from the server when run.

To get this to work within the build I had to add the old pcre.pro file
which gets output to a new makefile in the pcre directory which is
required when building qtbase.  Also, I had to remove the .dist extension
from pcre_charitables.c.dist.